Refactor tests to follow package structure
