+++ tests/value/conftest.py
@@ -1,0 +1,165 @@
+from collections import OrderedDict
+from typing import Dict
+
+import numpy as np
+import pytest
+from scipy.stats import spearmanr
+from sklearn.linear_model import LinearRegression
+from sklearn.pipeline import make_pipeline
+from sklearn.preprocessing import PolynomialFeatures
+
+from pydvl.utils import Dataset, Utility
+
+
+def polynomial(coefficients, x):
+    powers = np.arange(len(coefficients))
+    return np.power(x, np.tile(powers, (len(x), 1)).T).T @ coefficients
+
+
+@pytest.fixture(scope="function")
+def polynomial_dataset(coefficients: np.ndarray):
+    """Coefficients must be for monomials of increasing degree"""
+    from sklearn.utils import Bunch
+
+    x = np.arange(-1, 1, 0.05)
+    locs = polynomial(coefficients, x)
+    y = np.random.normal(loc=locs, scale=0.3)
+    db = Bunch()
+    db.data, db.target = x.reshape(-1, 1), y
+    poly = [f"{c} x^{i}" for i, c in enumerate(coefficients)]
+    poly = " + ".join(poly)
+    db.DESCR = f"$y \\sim N({poly}, 1)$"
+    db.feature_names = ["x"]
+    db.target_names = ["y"]
+    return Dataset.from_sklearn(data=db, train_size=0.15), coefficients
+
+
+@pytest.fixture(scope="function")
+def polynomial_pipeline(coefficients):
+    return make_pipeline(PolynomialFeatures(len(coefficients) - 1), LinearRegression())
+
+
+def dummy_utility(num_samples: int = 10):
+    from numpy import ndarray
+
+    from pydvl.utils import SupervisedModel
+
+    # Indices match values
+    x = np.arange(0, num_samples, 1).reshape(-1, 1)
+    nil = np.zeros_like(x)
+    data = Dataset(
+        x, nil, nil, nil, feature_names=["x"], target_names=["y"], description="dummy"
+    )
+
+    class DummyModel(SupervisedModel):
+        """Under this model each data point receives a score of index / max,
+        assuming that the values of training samples match their indices."""
+
+        def __init__(self, data: Dataset):
+            self.m = max(data.x_train)
+            self.utility = 0
+
+        def fit(self, x: ndarray, y: ndarray):
+            self.utility = np.sum(x) / self.m
+
+        def predict(self, x: ndarray) -> ndarray:
+            return x
+
+        def score(self, x: ndarray, y: ndarray) -> float:
+            return self.utility
+
+    return Utility(DummyModel(data), data, enable_cache=False)
+
+
+@pytest.fixture(scope="function")
+def analytic_shapley(num_samples):
+    """Scores are i/n, so v(i) = 1/n! Σ_π [U(S^π + {i}) - U(S^π)] = i/n"""
+    u = dummy_utility(num_samples)
+    exact_values = OrderedDict(
+        {i: i / float(max(u.data.x_train)) for i in u.data.indices}
+    )
+    return u, exact_values
+
+
+def check_total_value(
+    u: Utility, values: OrderedDict, rtol: float = 0.05, atol: float = 1e-6
+):
+    """Checks absolute distance between total and added values.
+    Shapley value is supposed to fulfill the total value axiom."""
+    total_utility = u(u.data.indices)
+    values = np.fromiter(values.values(), dtype=float, count=len(u.data))
+    # We could want relative tolerances here if we didn't have the range of
+    # the scorer.
+    assert np.isclose(values.sum(), total_utility, rtol=rtol, atol=atol)
+
+
+def check_exact(values: OrderedDict, exact_values: OrderedDict, atol: float = 1e-6):
+    """Compares ranks and values."""
+
+    k = list(values.keys())
+    ek = list(exact_values.keys())
+
+    assert np.all(k == ek), "Ranks do not match"
+
+    v = np.array(list(values.values()))
+    ev = np.array(list(exact_values.values()))
+
+    assert np.allclose(v, ev, atol=atol), f"{v} != {ev}"
+
+
+def check_values(
+    values: Dict,
+    exact_values: Dict,
+    rtol: float = 0.1,
+    atol: float = 1e-5,
+):
+    """Compares values in dictionaries.
+
+    Asserts that `|value - exact_value| < |exact_value| * rtol + atol` for
+    all pairs of `value`, `exact_value` with equal keys.
+
+    Note that this does not assume any ordering (despite values typically being
+    stored in an OrderedDict elsewhere.
+
+    :param values:
+    :param exact_values:
+    :param rtol: relative tolerance of elements in `values` with respect to
+        elements in `exact_values`. E.g. if rtol = 0.1, and atol = 0 we must
+        have |value - exact_value|/|exact_value| < 0.1 for every value
+    :param atol: absolute tolerance of elements in `values` with respect to
+        elements in `exact_values`. E.g. if atol = 0.1, and rtol = 0 we must
+        have |value - exact_value| < 0.1 for every value.
+    """
+    for key in values:
+        assert (
+            abs(values[key] - exact_values[key]) < abs(exact_values[key]) * rtol + atol
+        )
+
+
+def check_rank_correlation(
+    values: OrderedDict,
+    exact_values: OrderedDict,
+    k: int = None,
+    threshold: float = 0.9,
+):
+    """Checks that the indices of `values` and `exact_values` follow the same
+    order (by value), with some slack, using Spearman's correlation.
+
+    Runs an assertion for testing.
+
+    :param values: The values and indices to test
+    :param exact_values: The ground truth
+    :param k: Consider only these many, starting from the top.
+    :param threshold: minimal value for spearman correlation for the test to
+        succeed
+    """
+    # FIXME: estimate proper threshold for spearman
+    if k is not None:
+        raise NotImplementedError
+    else:
+        k = len(values)
+    ranks = np.array(list(values.keys())[:k])
+    ranks_exact = np.array(list(exact_values.keys())[:k])
+
+    correlation, pvalue = spearmanr(ranks, ranks_exact)
+    assert correlation >= threshold, f"{correlation} < {threshold}"
