+++ tests/value/shapley/test_knn.py
@@ -6,6 +6,7 @@ from sklearn.metrics import make_scorer
 from sklearn.neighbors import KNeighborsClassifier
 
 from pydvl.utils import Dataset, Utility, available_cpus
+from pydvl.value import SortOrder
 from pydvl.value.shapley.knn import knn_shapley
 from pydvl.value.shapley.naive import combinatorial_exact_shapley
 
@@ -44,7 +45,7 @@ def test_knn_montecarlo_match(seed):
     )
 
     # will check only matching top elements since the scoring functions are not exactly the same
-    exact_values.sort("desc")
+    exact_values.sort(SortOrder.Descending)
     top_knn = knn_values.indices[:2]
     top_exact = exact_values.indices[:4]
     assert np.all([k in top_exact for k in top_knn])
